HSEARCH-2726 Move Elasticsearch analyzer definition ITs to separate classes
This reduces code duplication in tests (since the tests are identical in
ES 2 and ES 5), and hopefully will make the addition of
normalizer-related tests easier.
